requirements.txt
fastapi
numpy
Pillow
requests
tensorflow-cpu
waitress
python-multipart